backend/requirements.txt
# Core
python-dotenv==1.0.0
pyyaml==6.0.1

# Data Processing
pandas==2.2.2
numpy==1.26.4
pyspark==3.5.3

# Market Data
yfinance==0.2.48

# Web Scraping
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0
selenium==4.27.1

# Sentiment Analysis
transformers==4.46.3
torch==2.5.1
vaderSentiment==3.3.2

# RAG & Vector DB
langchain==0.3.14
langchain-community==0.3.14
sentence-transformers==3.3.1
faiss-cpu==1.9.0
# chromadb==0.5.23  # COMMENTED OUT - requires C++ build tools

# Flask Backend
flask==3.1.0
flask-cors==5.0.0
flask-socketio==5.4.1

# Logging & Monitoring
loguru==0.7.3

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.2
tqdm==4.67.1

# Optional (for Ollama integration)
ollama==0.4.4